Refactoring on logging code
