+++ Backend/ServiceLayer/Services/AnalyticsService.cs
@@ -13,7 +13,7 @@ namespace ServiceLayer.Services
 {
     public class AnalyticsService : IAnalyticsService
     {
-        private const string ANALYTICS_SERVER_URL = "http://localhost:3000/";
+        private const string ANALYTICS_SERVER_URL = "https://julianjp.com/logging/";
 
         public HttpResponseMessage getAnalyticsData()
         {
