+++ Backend/ServiceLayer/Services/LoggingService.cs
@@ -15,7 +15,8 @@ namespace ServiceLayer.Services
 {
     public class LoggingService : ILoggingService
     {
-        private const string LOG_SERVER_URL = "https://julianjp.com/logging/";
+        //private const string LOG_SERVER_URL = "https://julianjp.com/logging/";
+        private const string LOG_SERVER_URL = "http://localhost:3000/";
         public static string LOGGER_API_SECRET = "CHRISTOPHER-123456-NIGHTWATCH-POINTMAP";
 
         public string GenerateSignature(string plaintext)
