+++ Loggingexc/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "loggingexc",
+  "version": "1.0.0",
+  "description": "Executable for managing logging in a collection",
+  "main": "app.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Chris Meyer",
+  "license": "MIT"
+}
